split risk analysis batch activity files
