+++ server/internal/background/activities/risk_analysis/analyze_batch.go
@@ -2,62 +2,30 @@ package risk_analysis
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log/slog"
-	"slices"
-	"strings"
-	"sync"
 	"time"
 
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/jackc/pgx/v5/pgxpool"
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/trace"
-	"go.temporal.io/sdk/activity"
 
 	riskv1 "github.com/speakeasy-api/gram/infra/gen/gram/risk/v1"
 	"github.com/speakeasy-api/gram/infra/pkg/gcp"
 	"github.com/speakeasy-api/gram/server/internal/attr"
-	"github.com/speakeasy-api/gram/server/internal/conv"
 	"github.com/speakeasy-api/gram/server/internal/feature"
 	"github.com/speakeasy-api/gram/server/internal/gitleaks"
-	"github.com/speakeasy-api/gram/server/internal/message"
 	"github.com/speakeasy-api/gram/server/internal/o11y"
-	"github.com/speakeasy-api/gram/server/internal/outbox"
-	"github.com/speakeasy-api/gram/server/internal/outbox/events"
 	"github.com/speakeasy-api/gram/server/internal/risk/celenv"
 	"github.com/speakeasy-api/gram/server/internal/risk/repo"
 	"github.com/speakeasy-api/gram/server/internal/shadowmcp"
-	"github.com/speakeasy-api/gram/server/internal/toolref"
 )
-
-// DescribeShadowMCP returns the canonical (rule_id, description) for an
-// unverified MCP tool call. Lives here because the writer
-// (scanMessageToolCalls) is the only caller.
-func DescribeShadowMCP(toolName string) (string, string) {
-	if toolName == "" {
-		return guard(RuleShadowMCP), "Detected an unverified MCP tool call."
-	}
-	return guard(RuleShadowMCP), fmt.Sprintf("Detected an unverified MCP tool call to %q.", toolName)
-}
-
-// DescribeDestructiveTool returns the canonical (rule_id, description) for
-// an MCP tool call whose resolved tool definition carries a destructive
-// annotation. Lives here because the writer
-// (scanMessageDestructiveToolCalls) is the only caller.
-func DescribeDestructiveTool(toolName string) (string, string) {
-	if toolName == "" {
-		return guard(RuleDestructiveTool), "Detected a call to a tool annotated as destructive by its MCP server."
-	}
-	return guard(RuleDestructiveTool), fmt.Sprintf("Detected a call to %q, which its MCP server annotates as destructive.", toolName)
-}
 
 // AnalyzeBatch scans a batch of messages against enabled detection sources
 // and writes the results back to the database.
@@ -129,7 +97,6 @@ type AnalyzeBatchArgs struct {
 	PromptInjectionRules []string
 	CustomRuleIds        []string
 }
-
 type AnalyzeBatchResult struct {
 	Processed int
 	Findings  int
@@ -273,7 +240,6 @@ func (a *AnalyzeBatch) Do(ctx context.Context, args AnalyzeBatchArgs) (_ *Analyz
 		Findings:  findingsCount,
 	}, nil
 }
-
 func (a *AnalyzeBatch) fetchContent(ctx context.Context, args AnalyzeBatchArgs) ([]repo.GetMessageContentBatchRow, error) {
 	ctx, fetchSpan := a.tracer.Start(ctx, "risk.fetchContent")
 	defer fetchSpan.End()
@@ -287,1055 +253,4 @@ func (a *AnalyzeBatch) fetchContent(ctx context.Context, args AnalyzeBatchArgs) 
 		return nil, fmt.Errorf("get message content batch: %w", err)
 	}
 	return messages, nil
-}
-
-func filterMessagesByMessageTypes(messages []repo.GetMessageContentBatchRow, messageTypes []string) []repo.GetMessageContentBatchRow {
-	filtered := make([]repo.GetMessageContentBatchRow, 0, len(messages))
-	for _, msg := range messages {
-		messageType, ok := messageRowMessageType(msg)
-		if !ok {
-			continue
-		}
-		if len(messageTypes) > 0 && !slices.Contains(messageTypes, messageType) {
-			continue
-		}
-		filtered = append(filtered, msg)
-	}
-	return filtered
-}
-
-func messageRowMessageType(msg repo.GetMessageContentBatchRow) (message.Type, bool) {
-	switch msg.Role {
-	case "user":
-		return message.User, true
-	case "tool":
-		return message.ToolResponse, true
-	case "assistant":
-		if len(msg.ToolCalls) > 0 {
-			return message.ToolRequest, true
-		}
-		return message.Assistant, true
-	default:
-		return "", false
-	}
-}
-
-// publishAckTimeout bounds how long we wait on a single shadow-mode Pub/Sub
-// publish ack while draining the issued futures. Kept well under the
-// AnalyzeBatch activity's 60s HeartbeatTimeout so a transient Pub/Sub stall
-// logs and moves on instead of blocking the goroutine past the heartbeat
-// window — otherwise Temporal cancels the activity and retries the whole
-// (expensive) scan. These publishes are fire-and-forget, so a timed-out ack is
-// only a warning.
-const publishAckTimeout = 10 * time.Second
-
-// drainPublishAcks waits on each shadow-mode publish future with a bounded
-// per-ack timeout, recording a heartbeat between acks so the worker keeps
-// signaling liveness throughout the drain. warnMsg is logged (with the ack
-// error) for any publish that fails or times out.
-func drainPublishAcks(ctx context.Context, logger *slog.Logger, warnMsg string, results []gcp.PublishResult) {
-	for _, res := range results {
-		waitCtx, cancel := context.WithTimeout(ctx, publishAckTimeout)
-		_, err := res.Get(waitCtx)
-		cancel()
-		if err != nil {
-			logger.WarnContext(ctx, warnMsg, attr.SlogError(err))
-		}
-		activity.RecordHeartbeat(ctx, "publish_ack")
-	}
-}
-
-// scan runs enabled scanners concurrently. Gitleaks (CPU-bound), presidio
-// (IO-bound), and prompt-injection (CPU-bound, regex-only) all run in
-// parallel — folding the cheap prompt-injection pass under presidio's
-// network wait keeps it free. All tool-call scanners (shadow_mcp,
-// destructive_tool, cli_destructive) run serially after the parallel scans
-// — shadow_mcp/destructive_tool make per-message DB calls; cli_destructive
-// is purely in-memory regex but kept in the same lane for consistency.
-func (a *AnalyzeBatch) scan(ctx context.Context, args AnalyzeBatchArgs, messages []repo.GetMessageContentBatchRow, customRules []CompiledCELRule, exclusions ExclusionSet, scopeExcluded []bool) ([][]Finding, error) {
-	ctx, scanSpan := a.tracer.Start(ctx, "risk.scanMessages")
-	defer scanSpan.End()
-	activity.RecordHeartbeat(ctx, 0)
-
-	n := len(messages)
-	contents := make([]string, n)
-	msgids := make([]uuid.UUID, n)
-	for i, msg := range messages {
-		contents[i] = msg.Content
-		msgids[i] = msg.ID
-	}
-
-	requestID, err := uuid.NewV7()
-	if err != nil {
-		return nil, fmt.Errorf("generate scan request id: %w", err)
-	}
-
-	gitleaksFindings := make([][]Finding, n)
-	presidioFindings := make([][]Finding, n)
-	shadowMCPFindings := make([][]Finding, n)
-	destructiveToolFindings := make([][]Finding, n)
-	cliDestructiveFindings := make([][]Finding, n)
-	promptInjectionFindings := make([][]Finding, n)
-	customFindings := make([][]Finding, n)
-
-	var wg sync.WaitGroup
-	var gitleaksErr error
-	var presidioErr error
-	var customErr error
-
-	if slices.Contains(args.Sources, "gitleaks") {
-		wg.Go(func() {
-			// Shadow-mode publish: fire a per-message GitleaksAnalysis so the
-			// streams subscriber re-runs the same scan and publishes findings.
-			// Issue every publish first so the Pub/Sub client can batch them;
-			// the returned futures are drained afterwards (see presidio below).
-			createdAt := time.Now().UTC().Format(time.RFC3339)
-			publishResults := make([]gcp.PublishResult, len(contents))
-			for i, content := range contents {
-				publishResults[i] = a.gitleaksPub.Publish(ctx, riskv1.GitleaksAnalysis_builder{
-					RequestId:         new(requestID.String()),
-					ChatMessageId:     new(msgids[i].String()),
-					ProjectId:         new(args.ProjectID.String()),
-					OrganizationId:    &args.OrganizationID,
-					RiskPolicyId:      new(args.RiskPolicyID.String()),
-					RiskPolicyVersion: &args.PolicyVersion,
-					CreatedAt:         &createdAt,
-
-					ReplyUrn: nil, // No reply support yet; fire and forget to exercise publish flow.
-					Content:  new(content),
-				}.Build())
-			}
-			drainPublishAcks(ctx, a.logger, "failed to publish gitleaks scan request", publishResults)
-
-			results, err := a.scanner.ScanBatchParallel(contents)
-			if err != nil {
-				gitleaksErr = err
-				return
-			}
-			for i, detections := range results {
-				gitleaksFindings[i] = fromDetections(detections)
-			}
-		})
-	}
-
-	if slices.Contains(args.Sources, "presidio") {
-		wg.Go(func() {
-			// Issue every publish first so the Pub/Sub client can batch them;
-			// the returned futures are drained afterwards. Calling Get inside
-			// the loop would block on each message's server ack before sending
-			// the next, defeating that batching.
-			createdAt := time.Now().UTC().Format(time.RFC3339)
-			publishResults := make([]gcp.PublishResult, len(contents))
-			for i, content := range contents {
-				publishResults[i] = a.presidioPub.Publish(ctx, riskv1.PresidioAnalysis_builder{
-					RequestId:         new(requestID.String()),
-					ChatMessageId:     new(msgids[i].String()),
-					ProjectId:         new(args.ProjectID.String()),
-					OrganizationId:    &args.OrganizationID,
-					RiskPolicyId:      new(args.RiskPolicyID.String()),
-					RiskPolicyVersion: &args.PolicyVersion,
-					CreatedAt:         &createdAt,
-
-					ReplyUrn: nil, // No reply support yet; fire and forget to exercise publish flow.
-					Content:  &content,
-					Entities: args.PresidioEntities,
-				}.Build())
-			}
-			drainPublishAcks(ctx, a.logger, "failed to publish presidio scan request", publishResults)
-
-			// PIIScanner may return partial results alongside an error;
-			// always consume results so successful per-text findings are
-			// preserved even when some HTTP calls failed.
-			results, err := a.piiScanner.AnalyzeBatch(ctx, contents, args.PresidioEntities, func() {
-				activity.RecordHeartbeat(ctx, "presidio")
-			})
-			if results != nil {
-				presidioFindings = results
-			}
-			if err != nil {
-				presidioErr = err
-				a.logger.WarnContext(ctx, "presidio scan returned errors, using partial results", attr.SlogError(err))
-				if a.metrics.presidioScanSkipped != nil {
-					a.metrics.presidioScanSkipped.Add(ctx, 1)
-				}
-			}
-		})
-	}
-
-	if slices.Contains(args.Sources, SourcePromptInjection) {
-		// Resolve the L1 engine opt-in once on the parent ctx (group-aware,
-		// before fan-out) and pass it down, mirroring the prompt-policies gate.
-		l1Enabled := a.projectFlagEnabled(ctx, args.OrganizationID, args.ProjectID, feature.FlagPromptInjectionUseClassifier)
-		// Build the structured messages the L1 judge reasons over (actor + tool
-		// attribution), aligned with contents — only when L1 will actually run.
-		var msgs []JudgeMessage
-		if l1Enabled {
-			msgs = make([]JudgeMessage, len(messages))
-			for i := range messages {
-				msgs[i] = a.judgeMessageForRow(ctx, messages[i])
-			}
-		}
-		wg.Go(func() {
-			results, err := a.piScanner.ScanBatch(ctx, contents, args.OrganizationID, args.ProjectID.String(), msgs, l1Enabled)
-			if err != nil {
-				a.logger.WarnContext(ctx, "prompt injection scan failed", attr.SlogError(err))
-				return
-			}
-			promptInjectionFindings = results
-			activity.RecordHeartbeat(ctx, "prompt_injection")
-		})
-	}
-
-	if len(customRules) > 0 {
-		wg.Go(func() {
-			eng := a.celEng
-			for i, msg := range messages {
-				findings, err := ScanCELRules(eng, a.customRuleMessageView(ctx, msg), customRules)
-				if err != nil {
-					customErr = err
-					return
-				}
-				customFindings[i] = findings
-			}
-			activity.RecordHeartbeat(ctx, "custom")
-		})
-	}
-
-	wg.Wait()
-
-	if gitleaksErr != nil {
-		scanSpan.SetStatus(codes.Error, gitleaksErr.Error())
-		return nil, fmt.Errorf("gitleaks scan batch: %w", gitleaksErr)
-	}
-
-	if customErr != nil {
-		scanSpan.SetStatus(codes.Error, customErr.Error())
-		return nil, fmt.Errorf("custom rule scan: %w", customErr)
-	}
-
-	// When the activity ctx was canceled (heartbeat timeout, parent
-	// workflow stop, etc.) report it as a scan-layer failure instead of
-	// letting control fall through to writeResults where db.Begin would
-	// surface a misleading "begin transaction: context canceled". Join
-	// the underlying Presidio error so the chain reflects the actual
-	// cause when Presidio degradation triggered the cancellation.
-	//
-	// Partial findings from gitleaks/promptInjection/presidio captured
-	// so far are intentionally discarded — Temporal will retry the
-	// activity (RetryPolicy.MaximumAttempts), and any partial write to
-	// the DB here would race the cancellation anyway.
-	if ctx.Err() != nil {
-		err := fmt.Errorf("scan canceled: %w", ctx.Err())
-		if presidioErr != nil {
-			err = errors.Join(err, fmt.Errorf("presidio: %w", presidioErr))
-		}
-		scanSpan.SetStatus(codes.Error, err.Error())
-		return nil, err
-	}
-
-	if slices.Contains(args.Sources, shadowmcp.SourceShadowMCP) {
-		shadowMCPFindings = a.scanShadowMCP(ctx, args.OrganizationID, args.ProjectID, messages)
-		activity.RecordHeartbeat(ctx, "shadow_mcp")
-	}
-
-	if slices.Contains(args.Sources, shadowmcp.SourceDestructiveTool) {
-		destructiveToolFindings = a.scanDestructiveToolAnnotations(ctx, args.OrganizationID, messages)
-		activity.RecordHeartbeat(ctx, "destructive_tool")
-	}
-
-	if slices.Contains(args.Sources, SourceCLIDestructive) {
-		cliDestructiveFindings = a.scanDestructiveCLICommands(ctx, messages)
-		activity.RecordHeartbeat(ctx, "cli_destructive")
-	}
-
-	merged := make([][]Finding, n)
-	for i := range n {
-		// The message is short-circuited — no findings recorded, including ones
-		// other detectors produced — when the policy's scope predicates put it out
-		// of scope or exempt it.
-		if scopeExcluded != nil && scopeExcluded[i] {
-			merged[i] = nil
-			continue
-		}
-		// Gitleaks findings come first so they take priority over presidio
-		// when both scanners match the same text region. Tool-call findings are
-		// non-overlapping with content scanners, so they pass through dedup.
-		combined := slices.Concat(gitleaksFindings[i], presidioFindings[i], shadowMCPFindings[i], destructiveToolFindings[i], cliDestructiveFindings[i], promptInjectionFindings[i], customFindings[i])
-		// Drop excluded findings before dedup so an excluded finding does not win
-		// the overlap and then vanish, leaving an overlapping (non-excluded)
-		// finding suppressed and the region unflagged.
-		if !exclusions.Empty() {
-			combined = exclusions.FilterFindings(combined)
-		}
-		merged[i] = dedup(combined)
-	}
-	return merged, nil
-}
-
-// scopeExclusions marks messages scope_include/scope_exempt take out of
-// scope (include does not match) or exempts. Returns nil when the policy has no
-// scope predicates, so the common path allocates nothing and callers skip
-// the per-message check.
-func (a *AnalyzeBatch) scopeExclusions(ctx context.Context, scope CompiledScope, messages []repo.GetMessageContentBatchRow) []bool {
-	if !scope.Active() {
-		return nil
-	}
-	excluded := make([]bool, len(messages))
-	for i, msg := range messages {
-		view := a.customRuleMessageView(ctx, msg)
-		excluded[i] = !scope.Includes(view) || scope.Exempts(view)
-	}
-	return excluded
-}
-
-// judgeConcurrency bounds the number of in-flight judge calls per batch. Judge
-// calls are LLM/network-bound (unlike the cheap in-memory/DB tool-call scanners,
-// which run serially because they are not the bottleneck), so a bounded pool
-// turns a ~batchSize×latency serial pass into roughly batchSize/judgeConcurrency
-// waves while staying well within OpenRouter rate/cost limits.
-const judgeConcurrency = 8
-
-// scanPromptJudge evaluates messages against a prompt_based policy's guardrail
-// prompt via the LLM judge, returning per-message findings aligned to messages.
-// Returns all-empty results when the judge is unavailable or the policy has no
-// prompt.
-//
-// The judge runs on every message passed in — the caller has already narrowed
-// them to the policy's message_types via filterMessagesByMessageTypes — so it
-// covers whatever types the policy declares. Messages are evaluated
-// concurrently with a bounded worker pool.
-func (a *AnalyzeBatch) scanPromptJudge(ctx context.Context, args AnalyzeBatchArgs, policy repo.RiskPolicy, messages []repo.GetMessageContentBatchRow, scopeExcluded []bool) [][]Finding {
-	out := make([][]Finding, len(messages))
-	cfg := ParseJudgeConfig(policy.ModelConfig)
-	if !a.projectFlagEnabled(ctx, args.OrganizationID, args.ProjectID, feature.FlagPromptPolicies) {
-		return out
-	}
-
-	indices := make([]int, 0, len(messages))
-	for i, msg := range messages {
-		// Skip messages the policy's scope predicates put out of scope or
-		// exempt — never pay for a judge call the policy wouldn't act on.
-		if scopeExcluded != nil && scopeExcluded[i] {
-			continue
-		}
-		if _, ok := messageRowMessageType(msg); ok {
-			indices = append(indices, i)
-		}
-	}
-	if len(indices) == 0 {
-		return out
-	}
-
-	if a.judge == nil || !policy.Prompt.Valid || strings.TrimSpace(policy.Prompt.String) == "" {
-		if !cfg.FailOpen {
-			finding := JudgeFinding(JudgeVerdict{Confidence: 0, Rationale: "Policy judge was unavailable; flagged by fail-closed policy."})
-			for _, idx := range indices {
-				out[idx] = []Finding{finding}
-			}
-		}
-		return out
-	}
-
-	// Distinct out[idx] writes per goroutine, so no shared-slice mutation.
-	// Heartbeat from the main goroutine between waves keeps the activity alive.
-	for start := 0; start < len(indices); start += judgeConcurrency {
-		end := min(start+judgeConcurrency, len(indices))
-		var wg sync.WaitGroup
-		for _, idx := range indices[start:end] {
-			wg.Add(1)
-			go func(idx int) {
-				defer wg.Done()
-				verdict := a.judge.Evaluate(ctx, JudgeInput{
-					OrgID:     args.OrganizationID,
-					ProjectID: args.ProjectID.String(),
-					Prompt:    policy.Prompt.String,
-					Message:   a.judgeMessageForRow(ctx, messages[idx]),
-					Config:    cfg,
-				})
-				if verdict != nil {
-					out[idx] = []Finding{JudgeFinding(*verdict)}
-				}
-			}(idx)
-		}
-		wg.Wait()
-		activity.RecordHeartbeat(ctx, "llm_judge", end)
-	}
-	return out
-}
-
-// projectFlagEnabled resolves a per-project PostHog feature flag against the
-// same org/project groups the dashboard registers so a group-targeted release
-// matches identically. A nil flag provider, or a failed slug or flag lookup,
-// degrades to disabled. Shared by the prompt-policies and prompt-injection-engine
-// gates.
-func (a *AnalyzeBatch) projectFlagEnabled(ctx context.Context, orgID string, projectID uuid.UUID, flag feature.Flag) bool {
-	if a.flags == nil {
-		return false
-	}
-	groups, err := repo.New(a.db).GetProjectFlagGroups(ctx, projectID)
-	if err != nil {
-		a.logger.WarnContext(ctx, "resolve project flag groups failed", attr.SlogError(err), attr.SlogOrganizationID(orgID), attr.SlogProjectID(projectID.String()))
-		return false
-	}
-	on, err := a.flags.IsFlagEnabled(ctx, flag, orgID, feature.OrgProjectGroups(groups.OrganizationSlug, groups.ProjectSlug))
-	if err != nil {
-		a.logger.WarnContext(ctx, "project flag check failed", attr.SlogError(err), attr.SlogOrganizationID(orgID))
-		return false
-	}
-	return on
-}
-
-// judgeMessageForRow builds the polymorphic judge payload for a stored message.
-// A single tool-call row surfaces the tool name + raw arguments; a multi-call
-// row surfaces each call with its own attribution so per-call MCP server and
-// function names reach the judge. A row with no parseable calls falls back to
-// the raw tool_calls JSON. Tool results and plain user/assistant messages carry
-// their content.
-func (a *AnalyzeBatch) judgeMessageForRow(ctx context.Context, msg repo.GetMessageContentBatchRow) JudgeMessage {
-	messageType, _ := messageRowMessageType(msg)
-	if messageType == message.ToolRequest {
-		calls := a.parseRecordedToolCalls(ctx, SourceLLMJudge, msg.ToolCalls)
-		switch len(calls) {
-		case 0:
-			// No parseable calls: hand over the raw array as opaque arguments.
-			return NewJudgeMessage(messageType, "", string(msg.ToolCalls))
-		case 1:
-			return NewJudgeMessage(messageType, calls[0].Function.Name, calls[0].Function.Arguments)
-		default:
-			judgeCalls := make([]JudgeToolCall, 0, len(calls))
-			for _, c := range calls {
-				// Skip malformed entries carrying neither a name nor arguments —
-				// they'd render as empty calls with no attribution to judge.
-				if c.Function.Name == "" && strings.TrimSpace(c.Function.Arguments) == "" {
-					continue
-				}
-				judgeCalls = append(judgeCalls, NewJudgeToolCall(c.Function.Name, c.Function.Arguments))
-			}
-			// If every entry was malformed, fall back to the raw array rather than
-			// hand the judge an empty tool_calls list.
-			if len(judgeCalls) == 0 {
-				return NewJudgeMessage(messageType, "", string(msg.ToolCalls))
-			}
-			return NewJudgeMessageForToolCalls(judgeCalls)
-		}
-	}
-	return NewJudgeMessage(messageType, "", msg.Content)
-}
-
-// customRuleMessageView builds the structured view the custom-rule engine
-// evaluates against: the message's text content, its type, and — for
-// tool-request messages — each recorded tool call with its MCP attribution.
-// Mirrors judgeMessageForRow so a rule sees the same per-call server/function
-// the judge does.
-func (a *AnalyzeBatch) customRuleMessageView(ctx context.Context, msg repo.GetMessageContentBatchRow) MessageView {
-	messageType, _ := messageRowMessageType(msg)
-	view := MessageView{Content: msg.Content, Type: messageType, Tools: nil}
-	if messageType != message.ToolRequest {
-		return view
-	}
-	for _, c := range a.parseRecordedToolCalls(ctx, SourceCustom, msg.ToolCalls) {
-		if c.Function.Name == "" && strings.TrimSpace(c.Function.Arguments) == "" {
-			continue
-		}
-		view.Tools = append(view.Tools, NewToolView(c.Function.Name, c.Function.Arguments))
-	}
-	return view
-}
-
-// customRulesForPolicy loads the custom rules a policy attaches as detectors
-// (risk_policies.custom_rule_ids) and compiles their CEL detection predicates.
-// Custom rules are pure detectors; message exemptions live in the policy's
-// scope_exempt, not in rules.
-func (a *AnalyzeBatch) customRulesForPolicy(ctx context.Context, projectID uuid.UUID, detectorIDs []string) ([]CompiledCELRule, error) {
-	if len(detectorIDs) == 0 {
-		return nil, nil
-	}
-
-	rules, err := repo.New(a.db).ListCustomDetectionRules(ctx, projectID)
-	if err != nil {
-		return nil, fmt.Errorf("list custom detection rules: %w", err)
-	}
-
-	detectors := make(map[string]struct{}, len(detectorIDs))
-	for _, id := range detectorIDs {
-		detectors[id] = struct{}{}
-	}
-
-	customRules := make([]CustomDetectionRule, 0, len(detectors))
-	for _, rule := range rules {
-		if _, ok := detectors[rule.RuleID]; !ok {
-			continue
-		}
-		customRules = append(customRules, CustomDetectionRule{
-			RuleID:        rule.RuleID,
-			Title:         rule.Title,
-			Description:   rule.Description,
-			DetectionExpr: rule.DetectionExpr.String,
-			Regex:         rule.Regex.String,
-		})
-	}
-
-	eng := a.celEng
-	compiled, err := CompileCELRules(eng, customRules)
-	if err != nil {
-		return nil, err
-	}
-	return compiled, nil
-}
-
-// scanShadowMCP validates each message's tool_calls against the shadow-MCP
-// guard. Messages without tool_calls (user prompts, assistant text, tool
-// results) are skipped. Each unsigned or mismatched call produces one Finding.
-//
-// Two-pass design: the first walk produces findings with placeholder match
-// values and accumulates the tool call IDs that triggered a deny; we then
-// do a single ClickHouse lookup for the `gram.mcp.match` attributes those
-// calls' PreToolUse hooks recorded, and patch them onto the findings. This
-// keeps shadow_mcp findings keyed by the same server identifier the live
-// hook saw (URL / stdio command), instead of a best-guess derivation from
-// the tool name alone.
-func (a *AnalyzeBatch) scanShadowMCP(ctx context.Context, orgID string, projectID uuid.UUID, messages []repo.GetMessageContentBatchRow) [][]Finding {
-	out := make([][]Finding, len(messages))
-	var deniedCallIDs []string
-	for i, msg := range messages {
-		if len(msg.ToolCalls) == 0 {
-			continue
-		}
-		findings, ids := a.scanMessageToolCalls(ctx, orgID, msg.ToolCalls)
-		out[i] = findings
-		deniedCallIDs = append(deniedCallIDs, ids...)
-	}
-
-	if len(deniedCallIDs) == 0 || a.mcpMatchLookup == nil {
-		return out
-	}
-
-	matches, err := a.mcpMatchLookup.LookupMCPMatchesByToolCallID(ctx, projectID, deniedCallIDs)
-	if err != nil {
-		// Non-fatal: the placeholder match (server prefix) is already on
-		// the findings, so the only consequence of a CH lookup failure is
-		// a slightly less precise allowlist key.
-		a.logger.WarnContext(ctx, "shadow_mcp scan: mcp match lookup failed; using server-prefix fallback",
-			attr.SlogError(err),
-		)
-		return out
-	}
-	for i := range out {
-		for j := range out[i] {
-			f := &out[i][j]
-			if f.Source != shadowmcp.SourceShadowMCP {
-				continue
-			}
-			if v, ok := matches[f.toolCallID]; ok && v != "" {
-				f.Match = v
-			}
-		}
-	}
-	return out
-}
-
-type recordedToolCall struct {
-	ID       string `json:"id"`
-	Function struct {
-		Name      string `json:"name"`
-		Arguments string `json:"arguments"`
-	} `json:"function"`
-}
-
-// MCPMatchLookup resolves a stored MCP tool call back to the server
-// identifier the hook saw at the time (`gram.mcp.match` on the
-// ClickHouse log). Returned map is keyed by tool call ID; missing
-// entries mean the hook never recorded a match for that call.
-type MCPMatchLookup interface {
-	LookupMCPMatchesByToolCallID(ctx context.Context, projectID uuid.UUID, toolCallIDs []string) (map[string]string, error)
-}
-
-func (a *AnalyzeBatch) parseRecordedToolCalls(ctx context.Context, source string, raw []byte) []recordedToolCall {
-	var calls []recordedToolCall
-	if err := json.Unmarshal(raw, &calls); err != nil {
-		a.logger.WarnContext(ctx, source+" scan: failed to parse tool_calls", attr.SlogError(err))
-		return nil
-	}
-	return calls
-}
-
-// scanMessageToolCalls iterates the tool_calls JSON array stored on a chat
-// message and runs the shadow_mcp validator against each call. The expected
-// payload mirrors what hooks/session_capture.go writes:
-// [{"id": "...", "type": "function", "function": {"name": "...", "arguments": "<json>"}}]
-//
-// Toolset lookups are served by the shared shadowmcp.Client cache so a batch
-// covering many calls from the same toolset only pays one DB round-trip.
-//
-// Returns the findings plus the list of tool call IDs that produced a deny —
-// scanShadowMCP collects these across messages so the resolved MCP match
-// (recorded by the hook on the ClickHouse log) can be patched onto each
-// finding in a single batched query.
-func (a *AnalyzeBatch) scanMessageToolCalls(ctx context.Context, orgID string, raw []byte) ([]Finding, []string) {
-	calls := a.parseRecordedToolCalls(ctx, shadowmcp.SourceShadowMCP, raw)
-
-	var findings []Finding
-	var deniedCallIDs []string
-	for _, call := range calls {
-		toolName := call.Function.Name
-		if toolName == "" {
-			continue
-		}
-		// Native (non-MCP) tools don't carry the x-gram-toolset-id property
-		// and are out of scope for shadow-MCP enforcement.
-		if !toolref.IsMCPToolName(toolName) {
-			continue
-		}
-		var toolInput any
-		if call.Function.Arguments != "" {
-			if err := json.Unmarshal([]byte(call.Function.Arguments), &toolInput); err != nil {
-				// Treat unparseable args as a missing toolset id.
-				toolInput = nil
-			}
-		}
-		bareName := toolref.MCPFunctionOf(toolName)
-		if a.shadowMCPClient == nil {
-			continue
-		}
-		_, denied := a.shadowMCPClient.ValidateToolsetCall(ctx, toolInput, bareName, orgID)
-		if !denied {
-			continue
-		}
-		// Placeholder match — the server-prefix portion of the tool name
-		// (e.g. "mise" from "mcp__mise__run_task"). scanShadowMCP's second
-		// pass replaces this with the authoritative `gram.mcp.match`
-		// attribute the hook recorded on the corresponding ClickHouse log,
-		// when one exists. The fallback keeps findings useful even if the
-		// CH lookup misses (no hook log yet, ClickHouse outage, ...).
-		match := toolref.MCPServerOf(toolName)
-		if match == "" {
-			match = toolName
-		}
-		ruleID, description := DescribeShadowMCP(toolName)
-		findings = append(findings, Finding{
-			Source:           shadowmcp.SourceShadowMCP,
-			RuleID:           ruleID,
-			Description:      description,
-			Match:            match,
-			StartPos:         0,
-			EndPos:           0,
-			Tags:             nil,
-			Confidence:       1.0,
-			DeadLetterReason: "",
-			toolCallID:       call.ID,
-			field:            "",
-			path:             "",
-		})
-		if call.ID != "" {
-			deniedCallIDs = append(deniedCallIDs, call.ID)
-		}
-	}
-	return findings, deniedCallIDs
-}
-
-// scanDestructiveToolAnnotations flags recorded Gram MCP tool calls whose
-// resolved tool definition carries a destructive annotation.
-func (a *AnalyzeBatch) scanDestructiveToolAnnotations(ctx context.Context, orgID string, messages []repo.GetMessageContentBatchRow) [][]Finding {
-	out := make([][]Finding, len(messages))
-	for i, msg := range messages {
-		if len(msg.ToolCalls) == 0 {
-			continue
-		}
-		out[i] = a.scanMessageDestructiveToolCalls(ctx, orgID, msg.ToolCalls)
-	}
-	return out
-}
-
-func (a *AnalyzeBatch) scanMessageDestructiveToolCalls(ctx context.Context, orgID string, raw []byte) []Finding {
-	if a.shadowMCPClient == nil {
-		return nil
-	}
-
-	calls := a.parseRecordedToolCalls(ctx, shadowmcp.SourceDestructiveTool, raw)
-
-	var findings []Finding
-	for _, call := range calls {
-		toolName := call.Function.Name
-		if toolName == "" || !toolref.IsMCPToolName(toolName) {
-			continue
-		}
-
-		var toolInput any
-		if call.Function.Arguments != "" {
-			if err := json.Unmarshal([]byte(call.Function.Arguments), &toolInput); err != nil {
-				continue
-			}
-		}
-
-		bareName := toolref.MCPFunctionOf(toolName)
-		resolved, ok := a.shadowMCPClient.ResolveToolsetCall(ctx, toolInput, bareName, orgID)
-		if !ok || resolved.Tool.Annotations == nil || resolved.Tool.Annotations.DestructiveHint == nil || !*resolved.Tool.Annotations.DestructiveHint {
-			continue
-		}
-
-		ruleID, description := DescribeDestructiveTool(resolved.ToolName)
-		findings = append(findings, Finding{
-			Source:           shadowmcp.SourceDestructiveTool,
-			RuleID:           ruleID,
-			Description:      description,
-			Match:            resolved.ToolName,
-			StartPos:         0,
-			EndPos:           0,
-			Tags:             nil,
-			Confidence:       1.0,
-			DeadLetterReason: "",
-			toolCallID:       "",
-			field:            "",
-			path:             "",
-		})
-	}
-	return findings
-}
-
-// scanDestructiveCLICommands flags recorded tool calls whose arguments contain
-// a curated destructive CLI pattern (rm -rf, git push --force, DROP TABLE,
-// ...). Unlike scanDestructiveToolAnnotations the trigger is content-driven,
-// so it applies to **every** tool call — native Bash / run_terminal_cmd as
-// well as MCP-routed calls whose args carry destructive content. The MCP
-// path can overlap with destructive_tool annotations; rule_id distinguishes
-// them and the dedup pass at the merge boundary is non-overlapping (start/end
-// positions are zero on tool-call findings).
-func (a *AnalyzeBatch) scanDestructiveCLICommands(ctx context.Context, messages []repo.GetMessageContentBatchRow) [][]Finding {
-	out := make([][]Finding, len(messages))
-	for i, msg := range messages {
-		if len(msg.ToolCalls) == 0 {
-			continue
-		}
-		out[i] = a.scanMessageDestructiveCLICalls(ctx, msg.ToolCalls)
-	}
-	return out
-}
-
-func (a *AnalyzeBatch) scanMessageDestructiveCLICalls(ctx context.Context, raw []byte) []Finding {
-	calls := a.parseRecordedToolCalls(ctx, SourceCLIDestructive, raw)
-
-	var findings []Finding
-	for _, call := range calls {
-		toolName := call.Function.Name
-		if toolName == "" {
-			continue
-		}
-
-		var toolInput any
-		if call.Function.Arguments != "" {
-			if err := json.Unmarshal([]byte(call.Function.Arguments), &toolInput); err != nil {
-				continue
-			}
-		}
-
-		matched, ok := scanForCLIDestructive(toolInput)
-		if !ok {
-			continue
-		}
-
-		ruleID, description := DescribeCLIDestructive(matched, toolName)
-		findings = append(findings, Finding{
-			Source:           SourceCLIDestructive,
-			RuleID:           ruleID,
-			Description:      description,
-			Match:            toolName,
-			StartPos:         0,
-			EndPos:           0,
-			Tags:             nil,
-			Confidence:       1.0,
-			DeadLetterReason: "",
-			toolCallID:       "",
-			field:            "",
-			path:             "",
-		})
-	}
-	return findings
-}
-
-// dedup removes findings that overlap the same text region. Earlier entries
-// in the slice win (gitleaks before presidio), so secrets take priority over PII.
-func dedup(findings []Finding) []Finding {
-	if len(findings) <= 1 {
-		return findings
-	}
-	var out []Finding
-	for _, f := range findings {
-		if overlapsAny(out, f) {
-			continue
-		}
-		out = append(out, f)
-	}
-	return out
-}
-
-func overlapsAny(kept []Finding, candidate Finding) bool {
-	for _, k := range kept {
-		if k.StartPos < candidate.EndPos && candidate.StartPos < k.EndPos {
-			return true
-		}
-	}
-	return false
-}
-
-func (a *AnalyzeBatch) buildRows(ctx context.Context, args AnalyzeBatchArgs, messages []repo.GetMessageContentBatchRow, batchFindings [][]Finding) ([]repo.InsertRiskResultsParams, int) {
-	var rows []repo.InsertRiskResultsParams
-	findingsCount := 0
-
-	for i, msg := range messages {
-		findings := batchFindings[i]
-
-		// Split dead-letter sentinels from real findings: DL markers do not
-		// contribute to findingsCount, do not feed the confidence histogram,
-		// and are stored with found=false + dead_letter_reason populated.
-		// They coexist with the per-message empty row so the existing
-		// "any row => analyzed" semantics in FetchUnanalyzedMessageIDs
-		// keep working until we add per-source dedup.
-		realFindings := findings[:0:0]
-		for _, f := range findings {
-			if f.DeadLetterReason != "" {
-				resultID, _ := uuid.NewV7()
-				rows = append(rows, deadLetterRow(resultID, args, msg.ID, f))
-				continue
-			}
-			realFindings = append(realFindings, f)
-		}
-
-		if len(realFindings) == 0 {
-			resultID, _ := uuid.NewV7()
-			rows = append(rows, emptyResultRow(resultID, args, msg.ID))
-			continue
-		}
-
-		for _, grp := range groupFindings(realFindings) {
-			f := grp.primary
-			findingsCount++
-			a.metrics.RecordFindingConfidence(ctx, args.OrganizationID, f.RuleID, f.Confidence)
-			resultID, _ := uuid.NewV7()
-			spansJSON, err := json.Marshal(grp.spans)
-			if err != nil {
-				// A span set that won't marshal is unexpected; fall back to no
-				// spans column rather than dropping the whole finding.
-				spansJSON = nil
-			}
-			rows = append(rows, repo.InsertRiskResultsParams{
-				ID:                resultID,
-				ProjectID:         args.ProjectID,
-				OrganizationID:    args.OrganizationID,
-				RiskPolicyID:      args.RiskPolicyID,
-				RiskPolicyVersion: args.PolicyVersion,
-				ChatMessageID:     msg.ID,
-				Source:            f.Source,
-				Found:             true,
-				RuleID:            pgtype.Text{String: f.RuleID, Valid: true},
-				Description:       pgtype.Text{String: f.Description, Valid: true},
-				Match:             pgtype.Text{String: f.Match, Valid: true},
-				StartPos:          pgtype.Int4{Int32: conv.SafeInt32(f.StartPos), Valid: true},
-				EndPos:            pgtype.Int4{Int32: conv.SafeInt32(f.EndPos), Valid: true},
-				Confidence:        pgtype.Float8{Float64: f.Confidence, Valid: true},
-				Tags:              f.Tags,
-				Spans:             spansJSON,
-				DeadLetterReason:  pgtype.Text{String: "", Valid: false},
-			})
-		}
-	}
-	return rows, findingsCount
-}
-
-// findingGroup is a single logical finding plus the spans attributed to it.
-type findingGroup struct {
-	primary Finding
-	spans   []FindingSpan
-}
-
-// groupFindings collapses findings that are spans of the same detection into one
-// group. Spans correlated to the same tool call (non-empty toolCallID) belong to
-// one finding — e.g. a custom rule matching both a tool's function name and its
-// arguments on the same call. Everything else (content/positional spans,
-// distinct detectors) stays its own finding, so genuinely distinct findings are
-// never merged. Order is preserved; the first span of a group is its primary.
-func groupFindings(findings []Finding) []findingGroup {
-	var order []string
-	groups := map[string]*findingGroup{}
-	uniq := 0
-	for _, f := range findings {
-		var key string
-		if f.toolCallID != "" {
-			key = f.Source + "\x00" + f.RuleID + "\x00" + f.toolCallID
-		} else {
-			key = fmt.Sprintf("u%d", uniq)
-			uniq++
-		}
-		g := groups[key]
-		if g == nil {
-			g = &findingGroup{primary: f, spans: nil}
-			groups[key] = g
-			order = append(order, key)
-		}
-		g.spans = append(g.spans, FindingSpan{
-			Match:    f.Match,
-			Field:    f.field,
-			Path:     f.path,
-			StartPos: f.StartPos,
-			EndPos:   f.EndPos,
-		})
-	}
-	out := make([]findingGroup, 0, len(order))
-	for _, k := range order {
-		out = append(out, *groups[k])
-	}
-	return out
-}
-
-func (a *AnalyzeBatch) writeResults(ctx context.Context, args AnalyzeBatchArgs, rows []repo.InsertRiskResultsParams) error {
-	ctx, writeSpan := a.tracer.Start(ctx, "risk.writeResults")
-	defer writeSpan.End()
-
-	rows = a.guardRuleIDs(ctx, rows)
-
-	tx, err := a.db.Begin(ctx)
-	if err != nil {
-		writeSpan.SetStatus(codes.Error, err.Error())
-		return fmt.Errorf("begin transaction: %w", err)
-	}
-	defer o11y.NoLogDefer(func() error { return tx.Rollback(ctx) })
-
-	txRepo := repo.New(a.db).WithTx(tx)
-
-	if _, err := txRepo.GetRiskPolicy(ctx, repo.GetRiskPolicyParams{
-		ID:        args.RiskPolicyID,
-		ProjectID: args.ProjectID,
-	}); err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			// Policy was soft-deleted while analysis was running — drop results.
-			writeSpan.SetAttributes(attribute.Bool("risk.policy_deleted", true))
-			a.logger.InfoContext(ctx, "risk policy deleted mid-analysis, dropping results",
-				attr.SlogRiskPolicyID(args.RiskPolicyID.String()),
-			)
-			return nil
-		}
-		writeSpan.SetStatus(codes.Error, err.Error())
-		return fmt.Errorf("re-check risk policy before writing results: %w", err)
-	}
-
-	if err := txRepo.DeleteRiskResultsForMessages(ctx, repo.DeleteRiskResultsForMessagesParams{
-		RiskPolicyID: args.RiskPolicyID,
-		ProjectID:    args.ProjectID,
-		MessageIds:   args.MessageIDs,
-	}); err != nil {
-		writeSpan.SetStatus(codes.Error, err.Error())
-		return fmt.Errorf("delete old results: %w", err)
-	}
-
-	if len(rows) > 0 {
-		if _, err := txRepo.InsertRiskResults(ctx, rows); err != nil {
-			writeSpan.SetStatus(codes.Error, err.Error())
-			return fmt.Errorf("insert risk results: %w", err)
-		}
-	}
-
-	now := time.Now()
-	var payloads []events.RiskFindingCreatedPayloadV1
-	for _, row := range rows {
-		if !row.Found || !row.RuleID.Valid {
-			continue
-		}
-		payloads = append(payloads, events.RiskFindingCreatedPayloadV1{
-			ID:                row.ID,
-			ProjectID:         row.ProjectID,
-			OrganizationID:    row.OrganizationID,
-			RiskPolicyID:      row.RiskPolicyID,
-			RiskPolicyVersion: row.RiskPolicyVersion,
-			ChatMessageID:     row.ChatMessageID,
-			RuleID:            row.RuleID.String,
-			Description:       row.Description.String,
-			Confidence:        row.Confidence.Float64,
-			Tags:              row.Tags,
-			CreatedAt:         now,
-		})
-	}
-	if len(payloads) > 0 {
-		if _, err := outbox.AppendBatch(ctx, tx, args.OrganizationID, events.RiskFindingCreatedV1, payloads); err != nil {
-			writeSpan.SetStatus(codes.Error, err.Error())
-			return fmt.Errorf("append risk findings to outbox: %w", err)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		writeSpan.SetStatus(codes.Error, err.Error())
-		return fmt.Errorf("commit results: %w", err)
-	}
-	return nil
-}
-
-// guardRuleIDs is the dev/test-only barrier before risk_results writes:
-// every row with a non-null rule_id must pass ValidateRuleID, otherwise
-// the batch panics so writer drift fails CI immediately. Production
-// passes rows through unchanged — dropping a row here would orphan the
-// message in the "no risk_results row = unanalyzed" semantics that
-// FetchUnanalyzedMessageIDs relies on (the message would be re-scanned
-// on every subsequent batch).
-//
-// Empty/null rule_ids are allowed; they represent the "analyzed, no
-// findings" sentinel row buildRows emits per message.
-func (a *AnalyzeBatch) guardRuleIDs(_ context.Context, rows []repo.InsertRiskResultsParams) []repo.InsertRiskResultsParams {
-	if !enforceRuleIDFormat {
-		return rows
-	}
-	for _, row := range rows {
-		if !row.RuleID.Valid || row.RuleID.String == "" {
-			continue
-		}
-		if err := ValidateRuleID(row.RuleID.String); err != nil {
-			panic(fmt.Sprintf("risk_analysis.writeResults: malformed rule_id %q from source %q: %v", row.RuleID.String, row.Source, err))
-		}
-	}
-	return rows
-}
-
-func emptyResultRow(id uuid.UUID, args AnalyzeBatchArgs, messageID uuid.UUID) repo.InsertRiskResultsParams {
-	return repo.InsertRiskResultsParams{
-		ID:                id,
-		ProjectID:         args.ProjectID,
-		OrganizationID:    args.OrganizationID,
-		RiskPolicyID:      args.RiskPolicyID,
-		RiskPolicyVersion: args.PolicyVersion,
-		ChatMessageID:     messageID,
-		Source:            "none",
-		Found:             false,
-		RuleID:            pgtype.Text{String: "", Valid: false},
-		Description:       pgtype.Text{String: "", Valid: false},
-		Match:             pgtype.Text{String: "", Valid: false},
-		StartPos:          pgtype.Int4{Int32: 0, Valid: false},
-		EndPos:            pgtype.Int4{Int32: 0, Valid: false},
-		Confidence:        pgtype.Float8{Float64: 0, Valid: false},
-		Tags:              nil,
-		Spans:             nil,
-		DeadLetterReason:  pgtype.Text{String: "", Valid: false},
-	}
-}
-
-// deadLetterRow materializes a Finding flagged with DeadLetterReason as a
-// risk_results row. found=false so the row never shows up in finding counts;
-// the rule_id/description carry the orchestrator's sentinel labels and
-// dead_letter_reason carries the underlying scanner error.
-func deadLetterRow(id uuid.UUID, args AnalyzeBatchArgs, messageID uuid.UUID, f Finding) repo.InsertRiskResultsParams {
-	return repo.InsertRiskResultsParams{
-		ID:                id,
-		ProjectID:         args.ProjectID,
-		OrganizationID:    args.OrganizationID,
-		RiskPolicyID:      args.RiskPolicyID,
-		RiskPolicyVersion: args.PolicyVersion,
-		ChatMessageID:     messageID,
-		Source:            f.Source,
-		Found:             false,
-		RuleID:            pgtype.Text{String: f.RuleID, Valid: f.RuleID != ""},
-		Description:       pgtype.Text{String: f.Description, Valid: f.Description != ""},
-		Match:             pgtype.Text{String: "", Valid: false},
-		StartPos:          pgtype.Int4{Int32: 0, Valid: false},
-		EndPos:            pgtype.Int4{Int32: 0, Valid: false},
-		Confidence:        pgtype.Float8{Float64: 0, Valid: false},
-		Tags:              nil,
-		Spans:             nil,
-		DeadLetterReason:  pgtype.Text{String: f.DeadLetterReason, Valid: true},
-	}
 }
